refactored the query related code
